+++ tests/integ/vpc_test_utils.py
@@ -127,13 +127,16 @@ def _create_vpc_resources(ec2_client, region, name):
     vpc_id = ec2_client.create_vpc(CidrBlock="10.0.0.0/16")["Vpc"]["VpcId"]
     print("created vpc: {}".format(vpc_id))
 
-    # sagemaker endpoints require subnets in at least 2 different AZs for vpc mode
+    availability_zone_name = ec2_client.describe_availability_zones()["AvailabilityZones"][0][
+        "ZoneName"
+    ]
+
     subnet_id_a = ec2_client.create_subnet(
-        CidrBlock="10.0.0.0/24", VpcId=vpc_id, AvailabilityZone=(region + "a")
+        CidrBlock="10.0.0.0/24", VpcId=vpc_id, AvailabilityZone=availability_zone_name
     )["Subnet"]["SubnetId"]
     print("created subnet: {}".format(subnet_id_a))
     subnet_id_b = ec2_client.create_subnet(
-        CidrBlock="10.0.1.0/24", VpcId=vpc_id, AvailabilityZone=(region + "b")
+        CidrBlock="10.0.1.0/24", VpcId=vpc_id, AvailabilityZone=availability_zone_name
     )["Subnet"]["SubnetId"]
     print("created subnet: {}".format(subnet_id_b))